rename directory "client" to "browser"
