+++ packages/frontend/web/components/MostViewed.tsx
@@ -12,7 +12,7 @@ import {
 import { screenReaderOnly } from '@guardian/pasteup/mixins';
 import { BigNumber } from '@guardian/guui';
 import { AsyncClientComponent } from './lib/AsyncClientComponent';
-import { reportError } from '@frontend/web/client/reportError';
+import { reportError } from '@frontend/web/browser/reportError';
 
 const container = css`
     border-top: 1px solid ${palette.neutral[86]};
